+++ examples/IRDispatcherDemo/IRCommandDispatcher.h
@@ -36,17 +36,16 @@
  */
 #define IR_COMMAND_FLAG_BLOCKING        0x00 // default - blocking command, repeat not accepted, only one command at a time. Stops an already running command.
 #define IR_COMMAND_FLAG_REPEATABLE      0x01 // repeat accepted
-#define IR_COMMAND_FLAG_NON_BLOCKING    0x02 // (Non blocking / non regular) (short) command that can be processed any time and may interrupt other IR commands - used for stop, set direction etc.
+#define IR_COMMAND_FLAG_NON_BLOCKING    0x02 // Non blocking (short) command that can be processed any time and may interrupt other IR commands - used for stop, set direction etc.
 #define IR_COMMAND_FLAG_REPEATABLE_NON_BLOCKING (IR_COMMAND_FLAG_REPEATABLE | IR_COMMAND_FLAG_NON_BLOCKING)
-/*
- * if this command is received, requestToStopReceived is set until call of next loop.
- * This stops ongoing commands which use:  RDispatcher.delayAndCheckForStop(100);  RETURN_IF_STOP;
- */
-#define IR_COMMAND_FLAG_IS_STOP_COMMAND 0x04 // sets requestToStopReceived (to stop other commands)
 
 // Basic mapping structure
 struct IRToCommandMappingStruct {
+#if defined(IR_COMMAND_HAS_MORE_THAN_8_BIT)
+    uint16_t IRCode;
+#else
     uint8_t IRCode;
+#endif
     uint8_t Flags;
     void (*CommandToCall)();
     const char *CommandString;
@@ -54,7 +53,11 @@ struct IRToCommandMappingStruct {
 
 struct IRDataForCommandDispatcherStruct {
     uint16_t address;           // to distinguish between multiple senders
+#if defined(IR_COMMAND_HAS_MORE_THAN_8_BIT)
     uint16_t command;
+#else
+    uint8_t command;
+#endif
     bool isRepeat;
     uint32_t MillisOfLastCode;  // millis() of last IR command -including repeats!- received - for timeouts etc.
     volatile bool isAvailable;  // flag for a polling interpreting function, that a new command has arrived. Is set true by library and set false by main loop.
@@ -63,8 +66,11 @@ struct IRDataForCommandDispatcherStruct {
 /*
  * Special codes (hopefully) not sent by the remote - otherwise please redefine it here
  */
-#define COMMAND_EMPTY       0xFE // code no command received
-#define COMMAND_INVALID     0xFF // code for command received, but not in mapping
+#if defined(IR_COMMAND_HAS_MORE_THAN_8_BIT)
+#define COMMAND_EMPTY       0xFFFF // code no command
+#else
+#define COMMAND_EMPTY       0xFF // code no command
+#endif
 
 #define RETURN_IF_STOP  if (IRDispatcher.requestToStopReceived) return
 #define BREAK_IF_STOP   if (IRDispatcher.requestToStopReceived) break
@@ -76,25 +82,36 @@ public:
 
     bool checkAndRunNonBlockingCommands();
     bool checkAndRunSuspendedBlockingCommands();
+#if defined(IR_COMMAND_HAS_MORE_THAN_8_BIT)
+    void setNextBlockingCommand(uint16_t aBlockingCommandToRunNext);
+#else
+    void setNextBlockingCommand(uint8_t aBlockingCommandToRunNext);
+#endif
     bool delayAndCheckForStop(uint16_t aDelayMillis);
 
     // The main dispatcher function
-    void checkAndCallCommand(bool aCallAlsoBlockingCommands);
+    void checkAndCallCommand(bool aCallBlockingCommandImmediately);
 
     void printIRCommandString(Print *aSerial);
     void setRequestToStopReceived(bool aRequestToStopReceived = true);
 
-    uint8_t currentBlockingCommandCalled = COMMAND_INVALID; // The code for the current called command
-    bool executingBlockingCommand = false;              // Lock for recursive calls of regular commands
+#if defined(IR_COMMAND_HAS_MORE_THAN_8_BIT)
+    uint16_t currentBlockingCommandCalled = COMMAND_EMPTY; // The code for the current called command
+    uint16_t lastBlockingCommandCalled = COMMAND_EMPTY; // The code for the last called command. Can be evaluated by main loop
+    uint16_t BlockingCommandToRunNext = COMMAND_EMPTY;  // Storage for command currently suspended to allow the current command to end, before it is called by main loop
+#else
+    uint8_t currentBlockingCommandCalled = COMMAND_EMPTY; // The code for the current called command
+    uint8_t lastBlockingCommandCalled = COMMAND_EMPTY;  // The code for the last called command. Can be evaluated by main loop
+    uint8_t BlockingCommandToRunNext = COMMAND_EMPTY;   // Storage for command currently suspended to allow the current command to end, before it is called by main loop
+#endif
     bool justCalledBlockingCommand = false;             // Flag that a blocking command was received and called - is set before call of command
-    uint8_t BlockingCommandToRunNext = COMMAND_INVALID; // Storage for command currently suspended to allow the current command to end, before it is called by main loop
     /*
      * Flag for running blocking commands to terminate. To check, you can use "if (IRDispatcher.requestToStopReceived) return;" (available as macro RETURN_IF_STOP).
      * Is reset by next IR command received. Can be reset by main loop, if command has stopped.
      */
     volatile bool requestToStopReceived;
     /*
-     * If we have a function, which want to interpret the IR codes by itself e.g. the calibrate function if QuadrupedControl then this flag must be true
+     * This flag must be true, if we have a function, which want to interpret the IR codes by itself e.g. the calibrate function of QuadrupedControl
      */
     bool doNotUseDispatcher = false;